refactor: Build the container per service, from that service's config
`src/container/index.ts` read `APP_CONFIG` from `@/services/handler/config`,
so an app with no service of that name did not compile. It is a factory now,
and each service passes its own config.

The factory is generic: `container.get("config")` has to keep the variables
the service adds on top of what the container reads.

`CONFIG_PROVIDER` and `CONFIG_KEY` move to `appConfigSchema` in
`domain/app-config.ts`. The container reads them whichever service built it,
so every service declares them.
