+++ templates/app/src/domain/app-config.ts
@@ -25,3 +25,25 @@ export const saleorMultiTenantAppConfig = saleorAppInstallations(appSettings);
 export type SaleorMultiTenantAppConfig = z.infer<
   typeof saleorMultiTenantAppConfig
 >;
+
+// Every service declares these, because the container they all build reads them.
+export const appConfigSchema = z.object({
+  CONFIG_PROVIDER: z
+    .enum(["edge", "file"])
+    .default("file")
+    .describe("Where the config of every installed Saleor is stored."),
+  CONFIG_KEY: z
+    .string()
+    .default("nimara-config")
+    .describe("Config provider key."),
+});
+
+/**
+ * What the container reads. Stated structurally rather than taken from one
+ * service's config, so a service that has no siblings still satisfies it.
+ */
+export type AppConfig = z.infer<typeof appConfigSchema> & {
+  ENVIRONMENT: string;
+  FETCH_TIMEOUT: number;
+  NAME: string;
+};
