+++ templates/app/src/services/handler/container.ts
@@ -1,0 +1,5 @@
+import { createAppContainer } from "@/container";
+
+import { APP_CONFIG } from "./config";
+
+export const container = createAppContainer(APP_CONFIG);
